+++ src/replication/mod.rs
@@ -4,5 +4,6 @@ pub mod backlog;
 pub mod graph_sync;
 pub mod handshake;
 pub mod master;
+pub mod reason_del;
 pub mod replica;
 pub mod state;